% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az_container_instance.R
\name{aci_ports}
\alias{aci_ports}
\alias{aci_creds}
\title{Utilities for specifying ACI configuration information}
\usage{
aci_ports(port = 80L, protocol = "TCP")

aci_creds(server, username, password)
}
\arguments{
\item{port, protocol}{For \code{aci_ports}, vectors of the port numbers and protocols to open for the instance.}

\item{server, username, password}{For \code{aci_creds}, the authentication details for a Docker registry.}
}
\description{
Utilities for specifying ACI configuration information
}
\details{
These are helper functions to be used in specifying the configuration for a container instance.
}
