% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kubernetes_cluster.R
\name{call_helm}
\alias{call_helm}
\title{Call the Helm commandline tool}
\usage{
call_helm(cmd = "", ...)
}
\arguments{
\item{cmd}{The Helm command line to execute.}

\item{...}{Other arguments to pass to \link{system2}.}
}
\value{
By default, the return code from the \code{helm} binary. The return value will have an added attribute \code{cmdline} that contains the command line. This makes it easier to construct scripts that can be run outside R.
}
\description{
Call the Helm commandline tool
}
\details{
This function calls the \code{helm} binary, which must be located in your search path. AzureContainers will search for the binary at package startup, and print a warning if it is not found.
}
\seealso{
\link{system2}, \link{call_docker}, \link{call_kubectl}

\link{kubernetes_cluster}

\href{https://kubernetes.io/docs/reference/kubectl/overview/}{Kubectl command line reference}
}
