% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_aci_methods.R
\name{delete_aci}
\alias{delete_aci}
\title{Delete an Azure Container Instance (ACI)}
\usage{
delete_aci(name, confirm=TRUE, wait=FALSE)
}
\arguments{
\item{name}{The name of the container instance.}

\item{confirm}{Whether to ask for confirmation before deleting.}

\item{wait}{Whether to wait until the deletion is complete.}
}
\value{
NULL on successful deletion.
}
\description{
Method for the \link[AzureRMR:az_resource_group]{AzureRMR::az_resource_group} class.
}
\seealso{
\link{create_aci}, \link{get_aci}

\link{az_container_instance}

\href{https://docs.microsoft.com/en-us/azure/container-instances/}{ACI documentation} and
\href{https://docs.microsoft.com/en-us/rest/api/container-instances/}{API reference}

\href{https://docs.docker.com/engine/reference/commandline/cli/}{Docker commandline reference}
}
