% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_acr}
\alias{is_acr}
\alias{is_aks}
\alias{is_aci}
\alias{is_docker_registry}
\alias{is_kubernetes_cluster}
\title{Utility functions to test whether an object is of the given class.}
\usage{
is_acr(object)

is_aks(object)

is_aci(object)

is_docker_registry(object)

is_kubernetes_cluster(object)
}
\arguments{
\item{object}{An R object}
}
\value{
TRUE or FALSE depending on whether the object is an R6 object of the specified class.
}
\description{
Utility functions to test whether an object is of the given class.
}
\details{
These functions are simple wrappers around \code{R6::is.R6} and \code{inherits}.
}
