% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az_kubernetes_service.R
\name{aks_pools}
\alias{aks_pools}
\title{Utility function for specifying Kubernetes agent pools}
\usage{
aks_pools(name, count, size = "Standard_DS2_v2", os = "Linux")
}
\arguments{
\item{name}{The name(s) of the pool(s).}

\item{count}{The number of nodes per pool.}

\item{size}{The VM type (size) to use for the pool.}

\item{os}{The operating system to use for the pool. Can be "Linux" or "Windows".}
}
\value{
A list of lists, suitable for passing to the \code{create_aks} constructor method.
}
\description{
Utility function for specifying Kubernetes agent pools
}
\details{
This is a convenience function to simplify the task of specifying the agent pool for a Kubernetes cluster. To specify multiple pools, provide vectors as input arguments; any scalar inputs will be replicated to match. Note that while Kubernetes in general is capable of having multiple pools per cluster, AKS currently only supports a single pool.
}
