% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_aks_methods.R
\name{create_aks}
\alias{create_aks}
\title{Create Azure Kubernetes Service (AKS)}
\usage{
create_aks(name, location = self$location,
           dns_prefix = name, kubernetes_version = "1.11.2",
           enable_rbac = FALSE, agent_pools = list(),
           properties = list(), ...)
}
\arguments{
\item{name}{The name of the Kubernetes service.}

\item{location}{The location/region in which to create the service. Defaults to this resource group's location.}

\item{dns_prefix}{The domain name prefix to use for the cluster endpoint. The actual domain name will start with this argument, followed by a string of pseudorandom characters.}

\item{kubernetes_version}{The Kubernetes version to use.}

\item{enable_rbac}{Whether to enable role-based access controls.}

\item{agent_pools}{A list of pool specifications. See 'Details'.}

\item{properties}{A named list of further Kubernetes-specific properties to pass to the initialization function.}

\item{...}{Other named arguments to pass to the initialization function.}
}
\value{
An object of class \code{az_kubernetes_service} representing the service.
}
\description{
Method for the \link[AzureRMR:az_resource_group]{AzureRMR::az_resource_group} class.
}
\details{
An AKS resource is a Kubernetes cluster hosted in Azure. See the \href{aks}{documentation for the resource} for more information. To work with the cluster (deploy images, define and start services, etc) see the \href{kubernetes_cluster}{documentation for the cluster endpoint}.

To specify the agent pools for the cluster, it is easiest to use the \link{aks_pools} function. This takes as arguments the name(s) of the pools, the number of nodes, the VM size(s) to use, and the operating system (Windows or Linux) to run on the VMs. Note that currently, AKS only supports one agent pool per cluster.
}
\seealso{
\link{get_aks}, \link{delete_aks}, \link{list_aks}, \link{aks_pools}

\link{az_kubernetes_service}

\link{kubernetes_cluster} for the cluster endpoint

\href{https://docs.microsoft.com/en-us/azure/aks/}{AKS documentation} and
\href{https://docs.microsoft.com/en-us/rest/api/aks/}{API reference}

\href{https://kubernetes.io/docs/reference/}{Kubernetes reference}
}
